requirements.txt
praw==7.8.1
boto3==1.28.0
apache-airflow==2.6.3
snowflake-connector-python==3.0.3
python-dotenv==1.0.0

